requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.8
djangorestframework==3.16.1
et_xmlfile==2.0.0
gunicorn==23.0.0
numpy==1.26.4
openpyxl==3.1.2
packaging==25.0
pandas==2.2.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.7.0
Pillow
